Move trades implementation into module
Leftover todo from previous PR.
